Allow tasks to be assigned to current user
Add logged in user as an option for assigning a task.  Decided
to add this as primary assignment option rather than as context role.
This necessitated adding the assignment type to the workflow (i.e user,
role, context role, logged in user) since previously we were inferring this
from the other properties set on the workflow (i.e existence of role id etc.)