Update to latest IntelliFlo.Platform dependencies
Encountered an issue with migrating templates where they had
automation triggers configured.  The problem was that the
TrustedClientAuthenticationDelegatingHandler from the old version of
the Platform was still using constructor injected
IIntelliFloClaimsPrincipal.  However when migrating we need to delegate
as different users and we do this by updating the
Thread.CurrentPrincipal.  Consequently TrustedAuth claims didn't
include UserId/TenantId which meant UserContext wasn't set in
destination API (i.e. eventmanagement/webhook).  This caused the
service to blow with an obscure NH Linq error (as wasn't checking for
null UserContext).