Improve exception when task fails to assign
Had an issue where due to testing migrations on my local build, a
template had not set the AssignedTo property when specifying 'Logged in
user'.  When creating instances of this template, the assigned user was
consequently set to the OwnerPartyid which was zero.  But as the RoleId
was null, this skipped throwing our FaultException, instead failed with
Bad Request when trying to create the task.