Optimise for workflow migration
Migration of instances involves recreating large numbers of instances
(~17K in UAT).  Running this previously ran into issues with memory
usage since each template was retained in memory as a running
WorkflowServiceHost service.

The primary optimisation was to track the loaded templates and instances
and unload templates where they no longer hold active instances.
Currently since we rely on WF's own Delay activity to handle delay
steps, we still need to maintain these templates in memory, but the
majority of instances will just be paused waiting for a callback on task
completion.  In future we could further optimise by replacing the Delay
activities with our own solution based around the Scheduler
microservice.  This would allow us to unload templates as soon as they
are done - in fact we may be able to switch to use WorkflowApplication
instead which would keep the instance/template alive only for that
particular call.

We also applied the following optimisations/fixes -

* Skip migration of previously migrated templates
* Update migration to handle delayed start
* Fix instance migration stopping after 500 rows
* Ignore unmigrated templates on startup

Fixes #6