Extract claims from token manually
We pass a bearer token to the workflow instance as a mechanism to
transfer the claims.  Previously we were calling
ITrustedAuthenticationScheme.Validate - which returned the claims -
however this has been altered to validate that the token hasn't expired.
Our token will most likely have expired, so we just need to extract the
claims manually.