Fix template creation when group is specified
We were firing TemplateGroupUpdated event before the template was saved
(i.e with a zero id).  This led to the handler blowing up because the
template could not be found.

Rather than refactoring at what point the event is fired instead we just
prevent firing of the event for new templates.  Only once the template
has been saved will the event fire.  For the requirements of that
handler this is sufficient, although it could be reviewed and refactored
later if necessary