Fix deletion of template
We had changed the underlying object for the Template domain object to
be a view - in order to allow us to efficiently retrieve related values
from the associated TemplateVersion.  However deleting a row from this
view therefore attempts to update more than one table and so fails.  As
a fix, add custom deletion statement to mapping