Fix issue where failing lead task caused crash
We were attempting to do continuation via a helper method that only
accepted an Action.  In the case of the lead task builder, the action
was defined as async.  However the code was not awaiting the results,
consequently when an exception was thrown it was unhandled and caused
the service to crash.

Fix by amending the extension method to be async and await the results.