Apply more consistent instance status label
When migrating to microservice, we had removed dependency on
StringValueAttribute which allowed us to format enum values with a
prettier name.  In doing this, we had renamed the InProgress enum value
to be Processing, in order to allow us to not have to do any processing
on it.  However this is misleading for the user.

Fix will split the enum value on capitalised letters allowing us to
return to the original "InProgress" (-> "In Progress") value.