Optimise query to retrieve delayed templates
On startup and every minute we retrieve the list of templates that are
stopped at a Delay activity.  We do this because these templates need to
remain loaded in the workflow service in order to process the activity
when the Delay activity ends so that the instance can continue.  The
previous query to retrieve this information was done purely using NH
LINQ queries and was inefficient.  Convert to a stored procedure which
returns list of populated TemplateDefinition objects instead.

As mentioned previously, the need for this may go away in future if we
move to using the Scheduler to call back to workflow when delays have
expired.  This would mean we don't need to rely on workflow Delay
activities and would only need to load instances whilst we are
processing them - we might also be able to move to use
WorkflowApplication instead of WorkflowServiceHost which would also
simplify things.