Resync step ids in instance history
When we migrated the templates we gave each step in a template a
specific Guid identifier.  These are also logged in the instance
history.  Previously these were just random generated guids for each step.

When we migrated we forgot to resync the history step ids with the new
ids assigned to that step.  Consequently when a step completes, it writes
the history against the new id, which fails to match the previous one -
leading to steps being left in a "In Progress" state from IO point of
view - although no behavioural change otherwise.