Correct logic for plan creation filter
The IsPreExisting property is poorly named.  In the UI, the selection is
New Business Only.  We invert this selection and set as IsPreExisting.
However when we convert to the filter we apply it directly so we end up
filtering for either pre-existing or not pre-existing.  But what we want
is to filter for not pre-existing or apply no filter.