Restart hosts when in Faulted state
Update WCF client calls to not use using clause in line with
recommendations (https://msdn.microsoft.com/en-us/library/aa355056
(v=vs.110).aspx)

When call is made to service in a Faulted state, close the service and
allow it to be restarted.

We are relying on call to start another instance in order to restart the
service - so this may lead to service getting left in Faulted state and
Delay activities not resuming, but will keep an eye on this.

When attempting to close service (i.e when purging), handle any
exception as a result of the service being in a Faulted state.