IP-1214 Fix instance filtering
Filtering was regressed by commit af151d6 which added NH query aliases.
NH Odata doesn't seem to play well when you have defined aliases and
calls querying on the aliased property fail.

As a workaround, moved to view and added TemplateTenantId property that
avoids the need to add alias.

This is possibly fixed in later version of Odata library, but we
currently are unable to upgrade as it would require an NH upgrade also.