IP-2614 Remove locking around increment/decrement of instance counter
Modify to use thread-safe counter instead
May still leave possible race condition where instance is created when
purge is about to close a service.  However service should
allow instance to progress and close the service gracefully - and there
is no longer possibility of deadlock when the instance tries to
decrement counter as service is unloaded.