IP-6974 Load only short-term delayed templates
+semver:patch

Previously we were loading all templates that had an delayed instance.
Occasionally we saw massive contention in the persistence database.  The
assumption is that this is exacerbated by having many loaded templates
especially as this same set is loaded on each of the workers.  This
short-term fix reduces the number of loaded templates that we maintain.